chore(refacto): add gameApi to centralize fetch
