+++ domains/game/GameScreen.jsx
@@ -4,6 +4,7 @@ import CONSTANTS from "../../const";
 import PropTypes from "prop-types";
 import Game from "./Game.jsx";
 import Waiting from "./Waiting.jsx";
+import gameApi from "../../api/GameApi.jsx";
 
 const isGameFull = (game) => {
     for (const player of game.players) {
@@ -23,13 +24,9 @@ const GameScreen = ({ route }) => {
     useEffect(() => {
         if (gameState === "pending") {
             setGameState("loading");
-            fetch(CONSTANTS.BASE_URL + "/game/" + gameId)
-                .then(async (res) => {
-                    const data = await res.json();
-                    if (!res.ok) {
-                        const error = (data && data.message) || res.status;
-                        return Promise.reject(error);
-                    }
+            gameApi
+                .getGame(gameId)
+                .then((data) => {
                     setGame(data);
                     setGameState("loaded");
                 })
@@ -42,14 +39,12 @@ const GameScreen = ({ route }) => {
     useEffect(() => {
         const updateGame = setInterval(function () {
             if (gameState === "loaded") {
-                fetch(CONSTANTS.BASE_URL + "/game/" + gameId)
-                    .then(async (res) => {
-                        const data = await res.json();
-                        if (!res.ok) {
-                            const error = (data && data.message) || res.status;
-                            return Promise.reject(error);
-                        }
+                console.log("updateGame", gameId);
+                gameApi
+                    .getGame(gameId)
+                    .then((data) => {
                         setGame(data);
+                        setGameState("loaded");
                     })
                     .catch((err) => {
                         setError(err);
